+++ api/urls.py
@@ -157,7 +157,7 @@ urlpatterns = [
         r'^v1/groups/(?P<pk>[0-9]+)$',
         v.ApiMember.as_view(
             model_class=RecipientGroup,
-            serializer_class=s.RecipientGroupSerializerMember,
+            serializer_class=s.RecipientGroupSerializer,
             permission_classes=(IsAuthenticated, CanSeeGroups)
         ),
         name='group'