+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Changed
 - Improve handling of scheduled messages
+- Update dependencies
 
 ## [v1.13.1]
 ### Fixed